requirements.txt
Pillow==11.1.0
rembg==2.0.61
streamlit==1.40.1
onnxruntime==1.19.2
